Move load_molecules from CLI to core openfe
